refactor: split processor package into smaller files
