+++ internal/reader/processor/processor_test.go
@@ -85,35 +85,6 @@ func TestAllowEntries(t *testing.T) {
 	}
 }
 
-func TestParseISO8601(t *testing.T) {
-	var scenarios = []struct {
-		duration string
-		expected time.Duration
-	}{
-		// Live streams and radio.
-		{"PT0M0S", 0},
-		// https://www.youtube.com/watch?v=HLrqNhgdiC0
-		{"PT6M20S", (6 * time.Minute) + (20 * time.Second)},
-		// https://www.youtube.com/watch?v=LZa5KKfqHtA
-		{"PT5M41S", (5 * time.Minute) + (41 * time.Second)},
-		// https://www.youtube.com/watch?v=yIxEEgEuhT4
-		{"PT51M52S", (51 * time.Minute) + (52 * time.Second)},
-		// https://www.youtube.com/watch?v=bpHf1XcoiFs
-		{"PT80M42S", (1 * time.Hour) + (20 * time.Minute) + (42 * time.Second)},
-	}
-
-	for _, tc := range scenarios {
-		result, err := parseISO8601(tc.duration)
-		if err != nil {
-			t.Errorf("Got an error when parsing %q: %v", tc.duration, err)
-		}
-
-		if tc.expected != result {
-			t.Errorf(`Unexpected result, got %v for duration %q`, result, tc.duration)
-		}
-	}
-}
-
 func TestIsRecentEntry(t *testing.T) {
 	parser := config.NewParser()
 	var err error
